Update preview generators packages
