+++ engine/src/main/java/org/terasology/rendering/nui/layers/mainMenu/preview/FacetLayerPreview.java
@@ -42,12 +42,13 @@ import org.terasology.world.generator.WorldGenerator;
 import org.terasology.world.viewer.color.ColorModels;
 import org.terasology.world.viewer.layers.FacetLayer;
 import org.terasology.world.viewer.layers.FacetLayers;
+import org.terasology.world.viewer.layers.Renders;
 
 import com.google.common.math.IntMath;
 
 /**
- * TODO Type description
- * @author Martin Steiger
+ * A world preview based on the facets that are generated by the {@link WorldGenerator}
+ * and a collection of {@link FacetLayer}, each annotated with {@link Renders} to display them.
  */
 public class FacetLayerPreview implements PreviewGenerator {
