+++ engine/src/main/java/org/terasology/world/viewer/color/ColorBlenderRgba.java
@@ -19,26 +19,31 @@ package org.terasology.world.viewer.color;
 
 
 /**
- * Blends RGBA colors using the ARGB color model
+ * Blends RGBA colors using the RGBA color model
  */
-class BlenderArgb implements Blender {
-
+class ColorBlenderRgba implements ColorBlender {
     @Override
     public int get(int src) {
-        // RGBA becomes ARGBA by a simple rotation
-        return Integer.rotateRight(src, Byte.SIZE);
+        return src;
     }
 
     @Override
-    public int add(int src1, int dst) {
-        int src = src1 >>> Byte.SIZE; // we ignore alpha
-        int mix = 0xFF000000;
-        mix |= Math.min(0x0000FF, (dst & 0x0000FF) + (src & 0x0000FF));
-        mix |= Math.min(0x00FF00, (dst & 0x00FF00) + (src & 0x00FF00));
-        mix |= Math.min(0xFF0000, (dst & 0xFF0000) + (src & 0xFF0000));
+    public int add(int src, int dst) {
+        int mix = 0x000000FF;
+        mix |= Math.min(0x0000FF00, (dst & 0x0000FF00) + (src & 0x0000FF00));
+        mix |= Math.min(0x00FF0000, (dst & 0x00FF0000) + (src & 0x00FF0000));
+
+        // casting to long avoids an integer overflow for the red channel
+        mix |= Math.min(0xFF000000L, (dst & 0xFF000000L) + (src & 0xFF000000L));
+
         return mix;
     }
 
+    /**
+     * @param src in RGBA
+     * @param dst in RGBA
+     * @return the blended color in RGBx
+     */
     @Override
     public int blend(int src, int dst) {
         int sr = (src >> 24) & 0xFF;
@@ -46,15 +51,15 @@ class BlenderArgb implements Blender {
         int sb = (src >> 8) & 0xFF;
         int a = src & 0xFF;
 
-        int dr = (dst >> 16) & 0xFF;
-        int dg = (dst >> 8) & 0xFF;
-        int db = dst & 0xFF;
+        int dr = (dst >> 24) & 0xFF;
+        int dg = (dst >> 16) & 0xFF;
+        int db = (dst >> 8) & 0xFF;
 
         int mb = (a * sb + (0xFF - a) * db) / 0xFF;
         int mg = (a * sg + (0xFF - a) * dg) / 0xFF;
         int mr = (a * sr + (0xFF - a) * dr) / 0xFF;
 
-        int mix = 0xFF000000 | mb | (mg << 8) | (mr << 16);
+        int mix = (mb << 8) | (mg << 16) | (mr << 24) | 0xFF;
         return mix;
     }
 }
